Moved settings from app/setup to application
Overriding of settings in tests is the source of state bleeding
between what should be atomic units. Application is in charge of
starting the application and its dependencies so loading the settings
have been moved there.

The setup module has been changed so that it is completely driven by
the config passed to create_app.

All unit tests pass. Integration tests are borked. More to come.